requirements.txt
# for LLM and RAG
torch==2.0.1
numpy==1.26.4
transformers>=4.38.2
langchain-community==0.3.17
huggingface-hub==0.28.1
langchain-huggingface==0.1.2
timm>=0.9.16
accelerate==1.3.0
sentencepiece==0.2.0
attrdict==2.0.1
einops==0.8.0
pypdf==5.3.0

# for FastAPI
fastapi==0.115.8
uvicorn==0.34.0
